Initial import to git
This is my first import of this script, gplv2, and readme.

This script takes a firmware xml file from motorola's uncompressed zip
file, and runs fastboot to flash it to your device.

Assuptions currently made are:
 * You have fastboot in your execution path envionment variable
 * You have uncompressed the firmware package and are in the same
   directory as the xml file and firmware binaries.
 * You have either copyied the moto-x-flash.py to the same directory or
   you have added this clone to your execution path environment
   variable.

This stuff my all change in the future, as the script needs to be
functionalized anyways. For now, I was able to run this and flash my
AT&T xt1058 16G with this script and the latest 4.2.2 firmware from:
http://sbf.droid-developers.org/ghost/list.php

Signed-off-by: Bryan Hundven <bryanhundven@gmail.com>